+++ CHANGELOG.md
@@ -2,11 +2,10 @@ Below are the noteworthy changes from each release.
 A more detailed list of changes is available in the corresponding milestones for each release in the Github issue tracker (https://github.com/googlefonts/fontbakery/milestones?state=closed).
 
 ## Upcoming release: 0.10.10 (2024-Jan-??)
-
 ### New checks
 #### On the Universal Profile
   - **EXPERIMENTAL - [com.google.fonts/check/varfont/family_axis_ranges]:** Check that family axis ranges are indentical. (issue #4445)
-  - **[com.google.fonts/check/tabular_kerning]:** Check that tabular numerals and symbols have no kerning
+  - **EXPERIMENTAL - [com.google.fonts/check/tabular_kerning]:** Check that tabular numerals and symbols have no kerning. (issue #4440)
 
 ### Noteworthy code-changes
   - The babelfont dependency has been dropped. (PR #4416)